requirements.txt
requests>=2.31.0
PyYAML>=6.0
openpyxl>=3.1.0
pandas>=2.0.0
python-dateutil>=2.8.0
click>=8.1.0
